requirements.txt
python-dotenv
pandas
numpy
pymongo
matplotlib
seaborn
certifi
pymongo[srv]
scikit-learn

#  -e .
